requirements.txt
django
tensorflow==2.10.0
numpy==1.24.3
psycopg2-binary